[refactor] Repository Query Result 패키지 경로 변경
