+++ news-api/src/main/kotlin/com/search/searchinfo/infrastructure/SearchInfoQueryRepository.kt
@@ -1,8 +1,8 @@
 package com.search.searchinfo.infrastructure
 
-import com.search.searchinfo.service.result.SearchInfoQueryResult
-import com.search.websearch.service.result.WebSearchResult
-import com.search.searchinfo.service.result.TopQueryResult
+import com.search.searchinfo.infrastructure.result.SearchInfoQueryResult
+import com.search.websearch.infrastructure.result.WebSearchResult
+import com.search.searchinfo.infrastructure.result.TopQueryResult
 import org.slf4j.LoggerFactory
 import org.springframework.jdbc.core.JdbcTemplate
 import org.springframework.jdbc.core.RowMapper
