+++ news-api/src/main/kotlin/com/search/websearch/infrastructure/KakaoWebRepositoryImpl.kt
@@ -5,8 +5,8 @@ import com.news.search.service.response.SearchQueryResponse
 import com.search.feign.KakaoClient
 import com.search.model.Document
 import com.search.websearch.service.port.WebRepository
-import com.search.websearch.service.result.WebSearchPageResult
-import com.search.websearch.service.result.WebSearchResult
+import com.search.websearch.infrastructure.result.WebSearchPageResult
+import com.search.websearch.infrastructure.result.WebSearchResult
 import org.springframework.stereotype.Repository
 
 @Repository
@@ -15,11 +15,11 @@ class KakaoWebRepositoryImpl(
 ) : WebRepository{
     override fun search(query: String, page: Int, size: Int): WebSearchPageResult<WebSearchResult> {
         val searchResult = kakaoClient.search(query,page, size)
-        val contents = searchResult.documents.map { toWebSearchResult(it) }.toList()
+        val contents = searchResult.documents.map { toWebSearchResult(it) }
         return WebSearchPageResult(page, size, searchResult.meta.totalCount, contents)
     }
 
-    private fun toWebSearchResult(document: Document): WebSearchResult{
+    private fun toWebSearchResult(document: Document): WebSearchResult {
         return WebSearchResult(
                 title = document.title,
                 link = document.url,
